requirements.txt
g4f>=0.3.0
edge-tts>=6.1.9
moviepy>=1.0.3
google-api-python-client>=2.100.0
google-auth>=2.23.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
